[update]: added login and register routes and FE UI

+++ api/src/routes/adminRoutes.ts
@@ -1,5 +1,9 @@
 // src/routes/adminRoutes.ts
 import { Router } from 'express'
+import { marked } from 'marked'
+import sanitizeHtml from 'sanitize-html'
+import { sanitizeOptions } from './publicRoutes'
+import { Post } from '../types/postTypes'
 
 const pool = require('../db')
 
@@ -8,14 +12,20 @@ export const adminRoutes = Router()
 // Create a new post
 adminRoutes.post('/posts', async (req, res) => {
   const { title, content, sub_title } = req.body
+
+  if (!req.user || !req.user.id) {
+    return res.status(403).send('Unauthorized access - user not identified')
+  }
+  const user_id = req.user.id
+
   if (!title || !content || !sub_title) {
     return res.status(400).send('Missing title or content')
   }
 
   try {
     const { rows } = await pool.query(
-      'INSERT INTO posts (title, content, sub_title) VALUES ($1, $2, $3) RETURNING *;',
-      [title, content, sub_title],
+      'INSERT INTO posts (title, content, sub_title, user_id) VALUES ($1, $2, $3, $4) RETURNING *;',
+      [title, content, sub_title, user_id],
     )
 
     const newPost = rows[0]
@@ -30,11 +40,20 @@ adminRoutes.post('/posts', async (req, res) => {
 adminRoutes.delete('/posts/:id', async (req, res) => {
   const postId = parseInt(req.params.id)
 
+  if (!req.user || !req.user.id) {
+    return res.status(403).send('Unauthorized access - user not identified')
+  }
+
   try {
     const fetchRes = await pool.query('SELECT * FROM posts WHERE id = $1;', [postId])
 
     if (fetchRes.rows.length === 0) {
       return res.status(404).send('Post not found')
+    }
+
+    // Check if the user ID associated with the post matches the authenticated user ID
+    if (fetchRes.rows[0].user_id !== req.user.id) {
+      return res.status(403).send('Unauthorized access - user not authorized to delete this post')
     }
 
     await pool.query('DELETE FROM posts WHERE id = $1;', [postId])
@@ -51,11 +70,20 @@ adminRoutes.put('/posts/:id', async (req, res) => {
   const postId = parseInt(req.params.id)
   const { title, content, sub_title } = req.body
 
+  if (!req.user || !req.user.id) {
+    return res.status(403).send('Unauthorized access - user not identified')
+  }
+
   try {
     const fetchRes = await pool.query('SELECT * FROM posts WHERE id = $1;', [postId])
 
     if (fetchRes.rows.length === 0) {
       return res.status(404).send('Post not found')
+    }
+
+    // Check if the user ID associated with the post matches the authenticated user ID
+    if (fetchRes.rows[0].user_id !== req.user.id) {
+      return res.status(403).send('Unauthorized access - user not authorized to update this post')
     }
 
     const { rows } = await pool.query(
@@ -67,5 +95,29 @@ adminRoutes.put('/posts/:id', async (req, res) => {
   } catch (error) {
     console.error(error)
     res.status(500).send('Error updating the post')
+  }
+})
+
+// Get all posts for the authenticated user
+adminRoutes.get('/my-posts', async (req, res) => {
+  if (!req.user || !req.user.id) {
+    return res.status(403).send('Unauthorized access - user not identified')
+  }
+  const user_id = req.user.id
+
+  try {
+    const { rows } = await pool.query(
+      'SELECT * FROM posts WHERE user_id = $1 ORDER BY created_at DESC;',
+      [user_id],
+    )
+    const processedPosts = rows.map((post: Post) => ({
+      ...post,
+      content: sanitizeHtml(marked(post.content) as string, sanitizeOptions),
+    }))
+
+    res.json(processedPosts)
+  } catch (error) {
+    console.error(error)
+    res.status(500).send('Error retrieving user posts')
   }
 })
